tfdb: Debug nodes inserter
EXPERIMENTAL: Insert special debug ops (e.g., DebugIdentity) to graph for
debugging. Currently, such ops need to take exactly one input and has the
string attribute "tensor_name" to indicate what tensor it watches.

For exapmle, before the node insertion, the graph may look like:

A:0 -----------1----------> B

wherein the output slot 0 of node A feeds as input to node B, through
edge 1.

After the node insertion, the graph becomes:

A:0 -----------1----------> B
    |                       ^
    |                       |
     ----2---> D ----3------

where D is the inserted debug node specified in the DebugNodeInserter's
constructor argument, 2 is an edge that sends A:0 to D, and 3 is a
control edge from D to B.

DebugIdentity is the simplest debugging paradigm, in which the debug
signal (i.e., output from D) equals the tensor itself. More sophisticated
debug ops can be used to transform of the tensor into other debug
signals.

A new field called "DebugTensorWatch" has been added to RunOptions
to support this.
Change: 126407058